Chemin multer + fs mis en place

+++ controllers/books.js
@@ -1,9 +1,10 @@
 const Book = require('../models/Book');
+const fs = require('fs');
 
 exports.createBook = (req, res, next) => {
     const bookData = JSON.parse(req.body.book)
     delete bookData._id;
-    delete bookData._userId;
+    delete bookData._userId; // revoir pourquoi suppression userId
     const book = new Book({
       ...bookData,
       userId: req.auth.userId,
@@ -26,22 +27,51 @@ exports.getOneBook = (req, res, next) => {
     .catch(error => res.status(404).json({ error }));
 };
 
-exports.getBestRating = (req, res, next) => {
-    Book.find({}).sort({averageRating: -1}).limit(3)
-    .then(books => res.status(200).json(books))
-    .catch(error => res.status(400).json({ error }));
-};
+// exports.getBestRating = (req, res, next) => {
+//     Book.find({}).sort({averageRating: -1}).limit(3)
+//     .then(books => res.status(200).json(books))
+//     .catch(error => res.status(400).json({ error }));
+// };
 
 exports.updateOneBook = (req, res, next) => {
-    Book.updateOne({ _id: req.params.id }, { ...req.body, _id: req.params.id })
-    .then(() => res.status(200).json({ message: 'Book modified !'}))
-    .catch(error => res.status(400).json({ error }));
+    const bookData = req.file ? {
+        ...JSON.parse(req.body.book),
+        imageUrl: `${req.protocol}://${req.get('host')}/images/${req.file.filename}`
+    } : {...req.body };
+
+    delete bookData._userId;
+    Book.findOne({_id: req.params.id})
+        .then ((book) => {
+            if (book.userId != req.auth.userId) {
+                res.status(401).json({ message:'Unauthorized' });
+            } else {
+                Book.updateOne({ _id: req.params.id }, { ...bookData, _id: req.params.id })
+                .then(() => res.status(200).json({ message: 'Book modified !'}))
+                .catch(error => res.status(400).json({ error }));
+            }
+        })
+        .catch((error) => {
+            res.status(400).json({ error });
+        });
 };
 
 exports.deleteOneBook = (req, res, next) => {
-    Book.deleteOne({ _id: req.params.id })
-    .then(() => res.status(200).json({ message: 'Book deleted !'}))
-    .catch(error => res.status(400).json({ error }));
+    Book.findOne({_id: req.params.id})
+        .then(book => {
+            if (book.userId != req.auth.userId) {
+                res.status(401).json({ message: 'Unauthorized' });
+            } else {
+                const filename = book.imageUrl.split('/images/')[1];
+                fs.unlink(`images/${filename}`, () => {
+                    Book.deleteOne({ _id: req.params.id })
+                    .then(() => res.status(200).json({ message: 'Book deleted !'}))
+                    .catch(error => res.status(400).json({ error }));
+                });
+            }
+        })
+        .catch((error) => {
+            res.status(500).json({ error });
+        });
 };
 
 // exports.createRating = (req, res, next) => {
